Update README.md
I'm a southern belle from ALABAMA the BEAUTIFUL, but currently residing in HOUSTON, TX. I have been trying for so long to teach myself how to become my own boss and make my own schedule as well as train and teach myself all the technological factors that are arising in this day and age. I am slowly going step by step and day by day just to attempt to make ends meat ... I am hoping and I believe that this opportunity will lead me into a beautiful field of work I have been searching for for so long. I am extremely excited and I am looking forward to learn and experience all the outs && ins of the technology world as well as the internet world!!!!!!!!! Knowledge is power so don't be shy; allow me to pick your brain so we can both succeed!!!!!!